Refactor to share circuit breaker with capi client
